Test : 컨테이너 삭제 테스트 ConatinerControllerTest에서 MainControllerTest로 이전
