+++ back/src/test/java/com/ogjg/back/container/controller/ContainerControllerTest.java
@@ -61,37 +61,6 @@ public class ContainerControllerTest extends ControllerTest {
         )).andExpect(status().isOk());
      }
 
-    @DisplayName("컨테이너 삭제")
-    @Test
-    public void deleteContainer() throws Exception {
-        //given
-        ContainerCreateRequest request = ContainerCreateRequest.builder()
-                .name("이회장")
-                .description("자바 연습 할거야")
-                .isPrivate(false)
-                .language("Java")
-                .build();
-
-        doNothing()
-                .when(containerService)
-                .deleteContainer(anyLong(), anyString());
-
-        // when
-        ResultActions result = this.mockMvc.perform(
-                delete("/api/containers/{containerId}", 1L)
-                        .accept(MediaType.APPLICATION_JSON)
-        );
-
-        // then
-        result.andDo(document("container/delete",
-                preprocessRequest(prettyPrint()),
-                preprocessResponse(prettyPrint()),
-                pathParameters(
-                        parameterWithName("containerId").description("컨테이너 ID")
-                )
-        )).andExpect(status().isOk());
-    }
-
     @DisplayName("컨테이너 이름 중복 체크")
     @Test
     public void checkDuplication() throws Exception {
